[elliot] Only set trust if different.
Fix for JENKINS-24427.

All connections use the same .p4trust file, which seems to get
clobbered sometimes. As addTrust writes to the file each time, I added
a check for whether we already have the current fingerprint, and only
add it if we don't.

Another solution would be to use per-connection trust files, through
setTrustFilePath().

#review-10093