More efficient listing of files.
Listing files for the Change reporting used `p4 files //…@=<id>`, this
now uses `p4 describe -s <id>` for changes and `p4 files //…@<id>` for
labels.

The `\\…@label` should be much more efficient as it only looks as the
db.label table not db.rev.  I can’t limit the scope to a workspace.
Often the client name uses ${var}, which are not present at this point
in the code.

JENKINS-25303