Fix change detection when using a moving label
If a build was pinned or built from a label, the label name was
recorded <buildChange>.  If the label is updated between builds then
comparing against an earlier build will give you the wrong result.

This fix stores the automatic label ‘Revision’ field and stores this in
<buildChange>.

NOTE: you will need to run a manual build to flush the label, so that
the last build has a change number.

JENKINS-25300