Behaviour: remove sync to head for unshelve step
Current workflow for review is to: sync the change, or label provided
then overrides this with a sync to head, before unshelving the content
— forcing the resolve step.

Removing the second sync allows users to determine the resolve point,
although they might not be able to submit the shelf without a second
resolve.

Old behaviour can be restored by fixing the review ‘change’ parameter
to ‘now’ a Perforce builtin revision.