Use latest change for all builds when polling
This fixes the case where multiple checkouts are present
in a pipeline for example and they are nested. Then using
the latest change instead of the first (as done previously)
ensures that continous builds are not triggered. This
does not fix the case where the checkouts are completely
disjoint.

JENKINS-40048