Tree walter for BranchSCMSource Events.
On a change-submit event use the change number and grab a submitted file. Walk up the path looking for a Jenkinsfile, then derive 'Project' path, 'Branch' name and 'Path' for the SCMHead/Revision.  This assumes that the Jenkinsfile is in the route of your projects.  Alternatively use the Swarm Event and pass your own 'Project', 'Branch' and 'Path'.