Cleanup step - use checkbox state and ignore Global option.
The global options for deleting the client is ignored and the cleanup checkbox state is used.  The Global option for deleting files still applies, but some what irrelevant as Jenkins can manage this in other steps.

Note that the RemoveClientTask is used for cleaning up temp clients in polling and checkout steps as wall as deleting a Jenkins Job.

JENKINS-47107