Always trigger a build if there's no previous one
When we can't determine the previously built change, always trigger a
build. Otherwise, Perforce triggers and Perforce polling will never
trigger a build until a manual build has been run.

It also seems that if a build fails before recording a last synced
change, it will "revert" to this "unbuilt" state, so further triggers
will not build the job.